restructure k6 project
restructure k6 project
