+++ loadtesting/k6/tests/utils/consts.js
@@ -1,0 +1,3 @@
+export const proxyUrl = __ENV.PROXY_URL;
+export const faucetUrl = __ENV.FAUCET_URL;
+export const networkId = parseInt(__ENV.NETWORK_ID);
